test(docx): move the resumed-numbering test to its own module
tests/test_backend_msword.py was 1496 lines and the new test pushed it to
1579, over the repository's 1500-line max-lines hook. Move the test into
tests/test_backend_msword_lists.py and restore test_backend_msword.py to
its state on main. No test logic changes.
